vkconfig: Refactor layers loading and override code
- Add layer override unit tests
- Prepare command line argument code to add this:
  - vkconfig layers --override ./configuration.json // apply configuration
  - vkconfig layers --surrender // delete override layers and settings files
