+++ vkconfig_core/override.h
@@ -17,3 +17,15 @@
  * Authors:
  * - Christophe Riccio <christophe@lunarg.com>
  */
+
+#pragma once
+
+#include "configuration.h"
+#include "environment.h"
+#include "application.h"
+
+// Create the VkLayer_override.json and vk_layer_settings.txt files to take over Vulkan layers from Vulkan applications
+bool OverrideLayers(const Environment& environment, const std::vector<Layer>& available_layers, const Configuration& configuration);
+
+// Remove the VkLayer_override.json and vk_layer_settings.txt files to return full control of the layers to the Vulkan applications
+bool SurrenderLayers(const Environment& environment);
